sources/azure: don't enforce nic ordering
The cloud-init generated netplan config persists between boots and the
names are applied early in boot by the netplan generator.  This may
result in the primary NIC being renamed away from "eth0", causing
cloud-init local to fail.

The ordering of NICs provided by IMDS may not match the order
enumerated by kernel, this will cause additional renaming thrashing
which tends to result in different types of failures with networkd
on Ubuntu 18.04, 20.04, and 22.04.

Stop setting device names until we can guarantee that:
(a) we can find the primary NIC regardless if it is eth0
(b) Ubuntu doesn't fail to bring up networking

Without ordering reliability, we cannot rely on checking the interface
device driver as named, so add a helper to reliably identify the most
appropriate matching.

LP #1958280

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>