azure/errors: introduce reportable errors
When provisioning failures occur an Azure, a generic description is
used in the report and ultimately is returned to the user.  To improve
the user experience, report details of the failure in a manner that is
parsable, readable and succinct.  The current approach is to use csv
with a custom delimiter ("|") and quote character ("'").  This format
may change in the future.

Gracefully handle reportable errors thrown while crawling metadata and
treat other exceptions as ReportableErrorUnhandledException.  Future
work will introduce more reportable errors to handle the expected cases.

Note that not all reportable errors are guaranteed to be fatal and that
and the datasource may recover after reporting a failure.  However, due
to the use of a variable timeout for provisioning, there are cases we
will report failure because of an issue likely to affect provisioning
within a reasonable timeframe (e.g. DHCP is very delayed, IMDS metadata
is unavailable, etc.).

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>