azure/imds: retry fetching metadata up to 300 seconds
Instead of a fixed number of retries, allow up to 5 minutes to fetch
metadata from IMDS.  The current approach allows for up to 11 attempts
depending on the path.  Given the timeout setting, this can vary from
~11 seconds up to ~32 seconds depending on whether or not read/connection
timeouts are encountered.

This will address the occasional delays and we'll be conducing testing
to help ensure the increase in latency does not significantly increase
provisioning timeouts.  If it does, we'll reduce the timeout.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>