azure/imds: remove limit for connection errors if route present
If there is a route to IMDS detected drop the limit for connection
errors as we know we've brought up the primary NIC and connectivity
should work.  In this case we'll continue until retry_deadline is up.

If there is no route to IMDS detected, such as Azure Stack, keep the
current limit of 11.

Fetching reprovision data behavior is unchanged.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>