fix(url_helper): fix TCP connection leak on readurl() retries
In scenarios where a lot of retries are expected, Ubuntu 24.04 fails
regularly with "Too many open files".  The `ulimit -n` shows the same
number of allowed open files in Ubuntu 20.04 (1024), but the connections
must be holding out longer on 24.04 for some reason. As the retry count
gets close to 1024 in readurl(), the open file limit is hit and
exceptions sprout up in a number of places.

The leak appears to occur because the same session is reused and all
prior failures are recorded in excps[] which in turn appears to keep
the connection open.

- create a new Session() for each connection attempt if session was not
passed in explicitly

- raise exceptions immediately when required rather than saving them to
do outside of the exception handler

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>